requirements.txt
streamlit==1.45.1
mysql-connector-python==8.3.0 